rename: component및 detail 폴더 정리
