+++ src/components/detail/MemberBox.tsx
