+++ src/components/detail/ScheduleCard.tsx
