+++ src/pages/detail/navComponent/Chat.tsx
